requirements.txt
# requirements.txt
# Все зависимости для проекта ML Trading Backtester

# ========== ОСНОВНЫЕ БИБЛИОТЕКИ ДЛЯ РАБОТЫ С ДАННЫМИ ==========
pandas>=1.5.0              # Работа с табличными данными (DataFrame)
numpy>=1.23.0              # Математические операции и работа с массивами
scipy>=1.9.0               # Научные вычисления, статистика

# ========== МАШИННОЕ ОБУЧЕНИЕ ==========
scikit-learn>=1.1.0        # Классические ML алгоритмы (Random Forest и др.)
xgboost>=1.6.0             # Градиентный бустинг XGBoost
lightgbm>=3.3.0            # Градиентный бустинг LightGBM

# ========== ГЛУБОКОЕ ОБУЧЕНИЕ ==========
tensorflow>=2.10.0         # Фреймворк для глубокого обучения (LSTM, GRU, Transformer)
keras>=2.10.0              # Высокоуровневый API для TensorFlow

# ========== ВИЗУАЛИЗАЦИЯ ==========
matplotlib>=3.5.0          # Построение графиков
seaborn>=0.12.0            # Красивые статистические графики
plotly>=5.10.0             # Интерактивные графики (опционально)

# ========== ТОРГОВЛЯ И ФИНАНСЫ ==========
MetaTrader5>=5.0.34        # Доступ к данным MetaTrader 5 (для загрузки котировок)

# ========== УТИЛИТЫ ==========
python-dotenv>=0.20.0      # Загрузка переменных окружения из .env файла
joblib>=1.2.0              # Сохранение и загрузка моделей
tqdm>=4.64.0               # Прогресс-бары для длительных операций
colorama>=0.4.5            # Цветной вывод в консоль (для Windows)

# ========== TELEGRAM БОТ (ОПЦИОНАЛЬНО) ==========
python-telegram-bot>=20.0  # Для отправки уведомлений в Telegram

# ========== ТЕСТИРОВАНИЕ ==========
pytest>=7.1.0              # Запуск тестов
pytest-cov>=3.0.0          # Оценка покрытия кода тестами

# ========== ДОПОЛНИТЕЛЬНЫЕ БИБЛИОТЕКИ (ОПЦИОНАЛЬНО) ==========
# pandas-ta>=0.3.14b0      # Технические индикаторы (альтернатива)
# ta-lib>=0.4.28           # Технические индикаторы (требует компиляции)
# statsmodels>=0.13.5      # Статистические модели (ARIMA и др.)
# prophet>=1.1.4           # Прогнозирование временных рядов от Facebook